+++ backend/token_api/views/action_set_view.py
@@ -14,7 +14,13 @@ from django.views.decorators.http import require_http_methods
 @require_http_methods(["POST"])
 def get_action_sets(request):
 
-    application_name = parse_arg(request, "application")
+    try:
+        application_name = parse_arg(request, "application")
+    except Exception:
+        return JsonResponse({"message": "Invalid json"})
+
+    if not application_name:
+        return JsonResponse({'message': "No application provided"})
 
     return JsonResponse({'message': serialize_action_set(ActionSet.objects.filter(application__application_name=application_name))})
 
@@ -23,8 +29,21 @@ def get_action_sets(request):
 @require_http_methods(["POST"])
 def get_action_set(request):
 
-    application_name = parse_arg(request, "application")
-    action_set_name = parse_arg(request, "action_set_name")
+    try:
+        application_name = parse_arg(request, "application")
+    except Exception:
+        return JsonResponse({"message": "Invalid json"})
+
+    if not application_name:
+        return JsonResponse({'message': "No application provided"})
+
+    try:
+        action_set_name = parse_arg(request, "action_set_name")
+    except Exception:
+        return JsonResponse({"message": "Invalid json"})
+
+    if not action_set_name:
+        return JsonResponse({'message': "No action_set_name provided"})
 
     try:
         account_policy = ActionSet.objects.get(application__application_name=application_name, action_set_name=action_set_name)
@@ -38,8 +57,15 @@ def get_action_set(request):
 @require_http_methods(["POST"])
 def update_action_set(request):
 
-    action_policy = parse_json(request)
-    deserializer_action_set([action_policy])
+    try:
+        action_set = parse_json(request)
+    except Exception:
+        return JsonResponse({"message": "Invalid json"})
+
+    try:
+        deserializer_action_set([action_set])
+    except Exception:
+        return JsonResponse({"message": "Invalid action set object"})
 
     return JsonResponse({"message": "Action Set updated"})
 
@@ -47,8 +73,21 @@ def update_action_set(request):
 @csrf_exempt
 @require_http_methods(["POST"])
 def remove_action_set(request):
-    application_name = parse_arg(request, "application")
-    action_set_name = parse_arg(request, "action_set_name")
+    try:
+        application_name = parse_arg(request, "application")
+    except Exception:
+        return JsonResponse({"message": "Invalid json"})
+
+    if not application_name:
+        return JsonResponse({'message': "No application provided"})
+
+    try:
+        action_set_name = parse_arg(request, "action_set_name")
+    except Exception:
+        return JsonResponse({"message": "Invalid json"})
+
+    if not action_set_name:
+        return JsonResponse({'message': "No action_set_name provided"})
 
     try:
         ActionSet.objects.get(application__application_name=application_name, action_set_name=action_set_name).delete()
